MASSIVE UPDATE
This update finalizes the BFL's Ultimate mode. Sorry, no noclip projectiles, BUT you get a temporary bubble shield during charge and AOE damage is emmited by the gun when charging.

New to this update also, a brand new REUSABLE bubble shield. To find this in the wild, pick up a blur sphere! Use with whatever key you may have bound to use. The bubble shield allows for 30 seconds of invulnerability, and takes 1 minute to recharge.

The Acrobat class also comes DEFAULT with the bubble shield now, so she'll be easier to play as.

Lots of new stuff. hope you all enjoy!