Assault Rifle!!! MERRY CHRISTMAS!!!
This update adds a whole new weapon that replaces the chaingun for now. Player classes have been rebalanced with their ammo counts, the Demon's Bane now has bullet chips, and a rework of the Annihilator made it function as I always have wanted.  I know there is some sprites to be made for a 3, 2, and 1 rocket clip, so when you get to the bottom of the pool, know I am working on that - AFTER CHRISTMAS!!!

Have a blessed Christmas!
-A Doomed Space Marine